+++ main/src/main/java/net/oneandone/stool/client/cli/Globals.java
@@ -21,24 +21,9 @@ import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.lang.reflect.InvocationTargetException;
 
 /** Basically a session factory */
 public class Globals {
-    public static Globals create(World world, FileNode itHome, String[] args) throws IOException {
-        Console console;
-        PrintWriter itOut;
-
-        if (itHome != null) {
-            itOut = new PrintWriter(itHome.join("client-output").newAppender(), true);
-            console = new Console(itOut, itOut, System.in);
-        } else {
-            console = Console.create();
-        }
-        return new Globals(itHome, args, console, world);
-    }
-
     private final FileNode itHome;
     private final String[] args;
     public final Console console;
@@ -59,25 +44,5 @@ public class Globals {
 
     public Server server() throws IOException {
         return net.oneandone.stool.server.cli.Globals.create(world, itHome, args).server();
-    }
-
-    //--
-
-    public int handleException(Throwable throwable) {
-        // TODO: inline should not throw InvocationTargetException ...
-        if (throwable instanceof InvocationTargetException) {
-            return handleException(((InvocationTargetException) throwable).getTargetException());
-        } else {
-            /* TODO:
-            if ((throwable instanceof RuntimeException) && (!(throwable instanceof ArgumentException))) {
-                try {
-                    session().reportException("RuntimeException", throwable);
-                } catch (IOException e) {
-                    console.error.println("failed to report runtine exception: " + e.getMessage());
-                    e.printStackTrace(console.verbose);
-                }
-            }*/
-            return console.handleException(throwable);
-        }
     }
 }